+++ src/main/java/net/frozenblock/wilderwild/client/renderer/entity/state/PenguinRenderState.java
@@ -31,4 +31,5 @@ public class PenguinRenderState extends LivingEntityRenderState {
 	public float slideProgress;
 	public AnimationState layDownAnimationState = new AnimationState();
 	public AnimationState standUpAnimationState = new AnimationState();
+	public AnimationState callAnimationState = new AnimationState();
 }