+++ src/main/java/net/frozenblock/wilderwild/worldgen/feature/configured/WWConfiguredFeatures.java
@@ -2304,11 +2304,14 @@ public final class WWConfiguredFeatures {
 				new NoiseProvider(
 					5050L,
 					new NormalNoise.NoiseParameters(0, 1D),
-					0.048833334F,
+					0.088833334F,
 					List.of(
+						WWBlocks.RED_HIBISCUS.defaultBlockState(),
+						WWBlocks.RED_HIBISCUS.defaultBlockState(),
 						WWBlocks.YELLOW_HIBISCUS.defaultBlockState(),
 						WWBlocks.WHITE_HIBISCUS.defaultBlockState(),
 						WWBlocks.PINK_HIBISCUS.defaultBlockState(),
+						WWBlocks.PURPLE_HIBISCUS.defaultBlockState(),
 						WWBlocks.PURPLE_HIBISCUS.defaultBlockState()
 					)
 				)
@@ -2324,7 +2327,7 @@ public final class WWConfiguredFeatures {
 
 		HIBISCUS_JUNGLE.makeAndSetHolder(Feature.FLOWER,
 			FeatureUtils.simpleRandomPatchConfiguration(
-				10,
+				16,
 				hibiscusNoise
 			)
 		);